cambio get id

+++ Back/SolRec/src/main/java/Fnt/SolRec/Model/Solicitud.java
@@ -29,7 +29,7 @@ public class Solicitud {
     /**
      * @return int return the id
      */
-    public int getId() {
+    public long getId() {
         return id;
     }
